@@ -29,17 +29,16 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "electron-log": "^2.2.11",
-    "electron-updater": "^2.16.1",
+    "electron-updater": "^2.16.3",
     "font-roboto-local": "^1.0.1",
     "highcharts": "^6.0.3",
-    "highcharts-more": "^0.1.2",
     "left-pad": "^1.2.0",
     "material-ui": "^0.19.4",
     "mdi": "^2.0.46",
-    "moment": "^2.19.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-google-maps": "^9.2.2",
+    "moment": "^2.19.2",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "react-google-maps": "^9.4.1",
     "react-highcharts": "^15.0.0",
     "react-markdown": "^3.0.2",
     "react-redux": "^5.0.6",
@@ -48,14 +47,16 @@
     "suncalc": "^1.8.0"
   },
   "devDependencies": {
-    "@types/node": "^8.0.46",
-    "@types/react": "^16.0.22",
+    "@types/googlemaps": "^3.30.0",
+    "@types/markerclustererplus": "^2.1.33",
+    "@types/node": "^8.0.53",
+    "@types/react": "^16.0.25",
     "@types/react-dom": "^16.0.3",
-    "@types/react-redux": "^5.0.12",
-    "electron": "^1.7.9",
-    "electron-builder": "^19.46.4",
+    "@types/react-redux": "^5.0.13",
+    "electron": "^1.8.1",
+    "electron-builder": "^19.46.6",
     "rimraf": "^2.6.2",
-    "typescript": "^2.5.3"
+    "typescript": "^2.6.1"
   },
   "build": {
     "productName": "Astroffers",
